Remove request options & move to did-method-key.
