+++ lib/didResolver.js
@@ -1,14 +1,11 @@
 /*!
  * Copyright (c) 2022 Digital Bazaar, Inc. All rights reserved.
  */
-import {
-  validateDidRequest,
-  validateRequestOptions
-} from './validators.js';
 import {didIo} from '@bedrock/did-io';
 import {DidResolutionResult} from './DidResolutionResult.js';
 import logger from './logger.js';
 import {parseDid} from '@digitalbazaar/did-io';
+import {validateDidRequest} from './validators.js';
 
 /**
  * Resolves a did after validating it.
@@ -17,8 +14,6 @@ import {parseDid} from '@digitalbazaar/did-io';
  * @param {string} options.did - The did or didUrl being resolved.
  * @param {Array<string>} options.supportedMethods - A list of
  *   did methods this resolver supports.
- * @param {Array<string>} options.publicKeyFormats - A list of
- *   public key formats this resolver supports.
  * @param {object} options.didOptions - Options passed to the resolver.
  *
  * @returns {object} - Returns a status code and the resolution data.
@@ -26,21 +21,15 @@ import {parseDid} from '@digitalbazaar/did-io';
 export const resolveDid = async ({
   did,
   supportedMethods,
-  publicKeyFormats,
-  didOptions
+  didOptions: rawOptions
 }) => {
   let error;
   try {
     const {method} = parseDid({did});
+    const didOptions = _convertDidOptions({didOptions: rawOptions});
     await validateDidRequest({method, supportedMethods, didOptions});
-    const didDocument = await didIo.get({did});
-    // FIXME didOptions should probably be in didIo
-    // in the future
-    validateRequestOptions({
-      didDocument,
-      didOptions,
-      publicKeyFormats
-    });
+    const didDocument = await didIo.get({did, options: didOptions});
+    // FIXME services are not handled here yet
     return {
       statusCode: 200,
       resolutionResult: new DidResolutionResult({did, didDocument})
@@ -96,4 +85,39 @@ function _errorCodeToStatusCode({code}) {
   const {statusCode = 500} = statuses.find(
     ({codes}) => codes.includes(code)) || {};
   return statusCode;
+}
+
+// takes in search params / url queries
+// and then converts properties that should be booleans
+// to booleans
+function _convertDidOptions({didOptions}) {
+  if(!didOptions) {
+    return;
+  }
+  if(didOptions.enableExperimentalPublicKeyTypes) {
+    didOptions.enableExperimentalPublicKeyTypes = _convertSearchParamToBoolean({
+      param: didOptions.enableExperimentalPublicKeyTypes
+    });
+  }
+  return didOptions;
+}
+
+/**
+ * Search Params are encoded as strings so this turns
+ * strings of value '0' or 'false' into false.
+ *
+ * @param {object} options - Options to use.
+ * @param {string} options.param - The param value.
+ *
+ * @returns {boolean} Returns a boolean.
+ */
+function _convertSearchParamToBoolean({param}) {
+  if(typeof param !== 'string') {
+    return Boolean(param);
+  }
+  const falsey = /0|false/i;
+  if(falsey.test(param)) {
+    return false;
+  }
+  return true;
 }
